chore: add stories
