+++ src/locales/de/translations.json
@@ -185,7 +185,7 @@
   "navigation.about-us": "Über Uns",
   "navigation.blog": "Blog",
   "navigation.contact": "Kontakt",
-  "blog.info": "Unsere Entwickler und Designer sprechen über die neuesten Trends und das Know-how rund um Tech, insbesondere über unsere neuesten Erkenntnisse und Einsichten zur Entwicklung guter Software für das Web.\n\n Der Blog ist nur auf Englisch verfügbar.",
+  "blog.info": "Unsere Entwickler und Designer sprechen über die neuesten Trends und das Know-how rund um Tech, insbesondere über unsere neuesten Erkenntnisse und Einsichten zur Entwicklung guter Software für das Web.",
   "blog.share": "Artikel teilen",
   "blog.follow": "Blog abonnieren",
   "blog.pagination": "Seite",
