+++ src/page-building/blog/blog-page.stories.mdx
@@ -1,0 +1,36 @@
+import { ArgsTable, Canvas, Meta, Story } from '@storybook/addon-docs';
+import { NotAvailableInGerman } from './not-avilable-in-german'; import { Posts } from './posts';
+
+<Meta title="Pages/Blog" />
+
+# Blog
+
+This is the blog overview page where people can browse the available blog posts.
+
+
+## Posts
+> The example can't show any Gatsby Image as this is not supported through storybook.
+
+In order to display the posts you can use the `<Posts/>` component and pass in a list of queries blog posts.
+
+<Canvas>
+  <Story name="Posts">
+    <Posts posts={[
+      {id: 'some-id1', frontmatter: {title: "My Blog Post", teaserText: 'Lorem ipsum dolar sit amet', path:' #'}},
+      {id: 'some-id2', frontmatter: {title: "My Blog Post", teaserText: 'Lorem ipsum dolar sit amet', path:' #'}},
+      {id: 'some-id3', frontmatter: {title: "My Blog Post", teaserText: 'Lorem ipsum dolar sit amet', path:' #'}}
+    ]}/>
+  </Story>
+</Canvas>
+
+<ArgsTable of={Posts} />
+
+## Not Available In German
+A message only displayed in the German version to redirect people to the English variant.
+This component is not covered by the translations as it's only valid in a single language.
+
+<Canvas>
+  <Story name="NotAvailableInGerman">
+    <NotAvailableInGerman/>
+  </Story>
+</Canvas>
